refactor(vm): move MRO/class-lookup reads into Registry methods (phase ② PR-B slice 1)
Phase ② PR-B is the read-side migration: factor the registry-read portions
of lookup / MRO / type-matching into methods on `Registry` itself, so the
read logic lives where the data lives rather than as multi-guard chains on
the Interpreter side.

MRO computation is the hottest registry read — every method dispatch walks
it. This slice moves it onto `impl Registry` (src/runtime/registry.rs):

- `Registry::compute_class_mro(&self, name, stack)` — pure C3 linearization
  over `self.classes`; recursion stays within the registry, no user-code
  re-entry.
- `Registry::class_mro(&mut self, name)` — builtin hierarchies + parameterized
  (`Foo[Bar]`) + cached `ClassDef::mro` + compute + cache write-back, all under
  a single write guard. Previously `Interpreter::class_mro` acquired up to five
  separate `registry()`/`registry_mut()` guards per call on the dispatch hot
  path; now one.
- `Registry::class_mro_cached(&self, name)` — read-only (cached or single).
- `Registry::class_has_method(&mut self, name, method)` — MRO-walk method probe.

Interpreter `class_mro`/`compute_class_mro`/`class_mro_readonly`/
`class_has_method` become thin delegating wrappers; the 65+ callers are
untouched.

Behavior-invariant structural refactor. build/clippy/make test green;
S12/S14 (class/role/MRO) roast green.

Notes for follow-up (recorded in docs/vm-registry-ownership.md): the VM has
no direct registry-field reads today (only 2 package_stubs writes) — its
registry access is all via Interpreter wrappers — so no VM registry handle is
added here. The type-matching composed-role transitive walk (two blocks in
type_matching.rs with divergent gate conditions) is deferred to a later slice
to keep this one behavior-invariant.
