+++ src/runtime/registry.rs
@@ -28,7 +28,7 @@ use std::collections::{HashMap, HashSet};
 
 use crate::ast::FunctionDef;
 use crate::symbol::Symbol;
-use crate::value::{EnumValue, Value};
+use crate::value::{EnumValue, RuntimeError, Value};
 
 use super::{ClassDef, RoleCandidateDef, RoleDef, SubsetDef};
 
@@ -118,4 +118,199 @@ pub(crate) struct Registry {
     pub(crate) proto_subs: HashSet<String>,
     /// `proto token`/`proto rule` declaration markers (existence set).
     pub(crate) proto_tokens: HashSet<String>,
+}
+
+/// Structural lookups over the declaration registry (PR-B: read-side migration).
+///
+/// These methods are the single source of truth for the *registry-read* portions
+/// of class lookup / MRO computation. They consult only registry fields (no
+/// re-entry into user-code execution and no Interpreter state), so they take a
+/// plain `&self` / `&mut self` on `Registry` and the caller holds exactly one
+/// guard for the whole operation — replacing the previous chains of separate
+/// `registry()` / `registry_mut()` acquisitions on the Interpreter side.
+impl Registry {
+    /// Compute the C3 linearization (method resolution order) for `class_name`
+    /// from the registered class hierarchy. Pure read over `self.classes` —
+    /// recursion stays within the registry. Does not consult or fill the cached
+    /// `ClassDef::mro` write side; that is done by [`Registry::class_mro`].
+    pub(crate) fn compute_class_mro(
+        &self,
+        class_name: &str,
+        stack: &mut Vec<String>,
+    ) -> Result<Vec<String>, RuntimeError> {
+        if stack.iter().any(|name| name == class_name) {
+            return Err(RuntimeError::new(format!(
+                "C3 MRO cycle detected at {}",
+                class_name
+            )));
+        }
+        if let Some(class_def) = self.classes.get(class_name)
+            && !class_def.mro.is_empty()
+        {
+            return Ok(class_def.mro.clone());
+        }
+        stack.push(class_name.to_string());
+        let explicit_parents = self
+            .classes
+            .get(class_name)
+            .map(|c| c.parents.clone())
+            .unwrap_or_default();
+        // If a user-defined class has no explicit parents, it implicitly
+        // inherits from Any (which in turn inherits from Mu).  This matches
+        // Raku's default class hierarchy.
+        let parents = if explicit_parents.is_empty() && self.classes.contains_key(class_name) {
+            vec!["Any".to_string()]
+        } else {
+            explicit_parents
+        };
+        let mut seqs: Vec<Vec<String>> = Vec::new();
+        for parent in &parents {
+            if self.classes.contains_key(parent) {
+                let mro = self.compute_class_mro(parent, stack)?;
+                seqs.push(mro);
+            } else if parent == "Any" {
+                // Any implicitly inherits from Mu
+                seqs.push(vec!["Any".to_string(), "Mu".to_string()]);
+            } else if parent == "Cool" {
+                seqs.push(vec![
+                    "Cool".to_string(),
+                    "Any".to_string(),
+                    "Mu".to_string(),
+                ]);
+            } else {
+                seqs.push(vec![parent.clone()]);
+            }
+        }
+        seqs.push(parents.clone());
+        let mut result = vec![class_name.to_string()];
+        while seqs.iter().any(|s| !s.is_empty()) {
+            let mut candidate = None;
+            for seq in &seqs {
+                if seq.is_empty() {
+                    continue;
+                }
+                let head = &seq[0];
+                let mut in_tail = false;
+                for other in &seqs {
+                    if other.len() > 1 && other[1..].contains(head) {
+                        in_tail = true;
+                        break;
+                    }
+                }
+                if !in_tail {
+                    candidate = Some(head.clone());
+                    break;
+                }
+            }
+            if let Some(head) = candidate {
+                result.push(head.clone());
+                for seq in seqs.iter_mut() {
+                    if !seq.is_empty() && seq[0] == head {
+                        seq.remove(0);
+                    }
+                }
+            } else {
+                stack.pop();
+                return Err(RuntimeError::new(format!(
+                    "Inconsistent class hierarchy for {}",
+                    class_name
+                )));
+            }
+        }
+        stack.pop();
+        Ok(result)
+    }
+
+    /// Resolve the MRO for `class_name`, returning the cached `ClassDef::mro`
+    /// when present, the hardcoded hierarchy for built-in types that are not
+    /// user-defined classes, and otherwise computing + caching via
+    /// [`Registry::compute_class_mro`]. Single write guard for the whole op.
+    pub(crate) fn class_mro(&mut self, class_name: &str) -> Vec<String> {
+        // Built-in type hierarchies for types that are not user-defined classes
+        if !self.classes.contains_key(class_name) {
+            let builtin_mro: Option<Vec<&str>> = match class_name {
+                "Match" => Some(vec!["Match", "Capture", "Cool", "Any", "Mu"]),
+                "Capture" => Some(vec!["Capture", "Any", "Mu"]),
+                "Distribution::Path" => Some(vec!["Distribution::Path", "Distribution", "Any", "Mu"]),
+                "Distribution::Hash" => Some(vec!["Distribution::Hash", "Distribution", "Any", "Mu"]),
+                "Distribution::Installation" => Some(vec![
+                    "Distribution::Installation",
+                    "Distribution",
+                    "Any",
+                    "Mu",
+                ]),
+                "CompUnit::DependencySpecification" => {
+                    Some(vec!["CompUnit::DependencySpecification", "Any", "Mu"])
+                }
+                "CompUnit::Repository::FileSystem" => Some(vec![
+                    "CompUnit::Repository::FileSystem",
+                    "CompUnit::Repository",
+                    "Any",
+                    "Mu",
+                ]),
+                "CompUnit::Repository::Installation" => Some(vec![
+                    "CompUnit::Repository::Installation",
+                    "CompUnit::Repository::Installable",
+                    "CompUnit::Repository::Locally",
+                    "CompUnit::Repository",
+                    "Any",
+                    "Mu",
+                ]),
+                _ => None,
+            };
+            if let Some(mro) = builtin_mro {
+                return mro.into_iter().map(String::from).collect();
+            }
+        }
+        if !self.classes.contains_key(class_name)
+            && let Some((base, _)) = class_name.split_once('[')
+            && class_name.ends_with(']')
+            && self.classes.contains_key(base)
+        {
+            let mut mro = vec![class_name.to_string()];
+            mro.extend(self.class_mro(base));
+            return mro;
+        }
+        if let Some(class_def) = self.classes.get(class_name)
+            && !class_def.mro.is_empty()
+        {
+            return class_def.mro.clone();
+        }
+        let mut stack = Vec::new();
+        match self.compute_class_mro(class_name, &mut stack) {
+            Ok(mro) => {
+                if let Some(class_def) = self.classes.get_mut(class_name) {
+                    class_def.mro = mro.clone();
+                }
+                mro
+            }
+            Err(_) => vec![class_name.to_string()],
+        }
+    }
+
+    /// Read-only MRO lookup: the cached `ClassDef::mro` when present, otherwise a
+    /// single-element MRO `[class_name]`. Returns `None` when the class is not
+    /// registered. Used by non-`&mut` helpers that must not trigger computation.
+    pub(crate) fn class_mro_cached(&self, class_name: &str) -> Option<Vec<String>> {
+        let class_def = self.classes.get(class_name)?;
+        if !class_def.mro.is_empty() {
+            return Some(class_def.mro.clone());
+        }
+        Some(vec![class_name.to_string()])
+    }
+
+    /// Whether `class_name` (or any class in its MRO) defines `method_name`
+    /// either as a user method or a native method. Pure registry MRO walk.
+    pub(crate) fn class_has_method(&mut self, class_name: &str, method_name: &str) -> bool {
+        let mro = self.class_mro(class_name);
+        for cn in mro {
+            if let Some(class_def) = self.classes.get(&cn)
+                && (class_def.methods.contains_key(method_name)
+                    || class_def.native_methods.contains(method_name))
+            {
+                return true;
+            }
+        }
+        false
+    }
 }
