+++ src/runtime/resolution.rs
@@ -908,71 +908,10 @@ impl Interpreter {
             .is_some_and(|caller| caller == owner_class)
     }
 
+    /// Resolve the MRO for `class_name`. Delegates to [`Registry::class_mro`],
+    /// which holds a single write guard for the whole compute-and-cache op.
     pub(super) fn class_mro(&mut self, class_name: &str) -> Vec<String> {
-        // Built-in type hierarchies for types that are not user-defined classes
-        if !self.registry().classes.contains_key(class_name) {
-            let builtin_mro: Option<Vec<&str>> = match class_name {
-                "Match" => Some(vec!["Match", "Capture", "Cool", "Any", "Mu"]),
-                "Capture" => Some(vec!["Capture", "Any", "Mu"]),
-                "Distribution::Path" => {
-                    Some(vec!["Distribution::Path", "Distribution", "Any", "Mu"])
-                }
-                "Distribution::Hash" => {
-                    Some(vec!["Distribution::Hash", "Distribution", "Any", "Mu"])
-                }
-                "Distribution::Installation" => Some(vec![
-                    "Distribution::Installation",
-                    "Distribution",
-                    "Any",
-                    "Mu",
-                ]),
-                "CompUnit::DependencySpecification" => {
-                    Some(vec!["CompUnit::DependencySpecification", "Any", "Mu"])
-                }
-                "CompUnit::Repository::FileSystem" => Some(vec![
-                    "CompUnit::Repository::FileSystem",
-                    "CompUnit::Repository",
-                    "Any",
-                    "Mu",
-                ]),
-                "CompUnit::Repository::Installation" => Some(vec![
-                    "CompUnit::Repository::Installation",
-                    "CompUnit::Repository::Installable",
-                    "CompUnit::Repository::Locally",
-                    "CompUnit::Repository",
-                    "Any",
-                    "Mu",
-                ]),
-                _ => None,
-            };
-            if let Some(mro) = builtin_mro {
-                return mro.into_iter().map(String::from).collect();
-            }
-        }
-        if !self.registry().classes.contains_key(class_name)
-            && let Some((base, _)) = class_name.split_once('[')
-            && class_name.ends_with(']')
-            && self.registry().classes.contains_key(base)
-        {
-            let mut mro = vec![class_name.to_string()];
-            mro.extend(self.class_mro(base));
-            return mro;
-        }
-        if let Some(class_def) = self.registry().classes.get(class_name)
-            && !class_def.mro.is_empty()
-        {
-            return class_def.mro.clone();
-        }
-        let mut stack = Vec::new();
-        match self.compute_class_mro(class_name, &mut stack) {
-            Ok(mro) => {
-                if let Some(class_def) = self.registry_mut().classes.get_mut(class_name) {
-                    class_def.mro = mro.clone();
-                }
-                mro
-            }
-            Err(_) => vec![class_name.to_string()],
-        }
+        self.registry_mut().class_mro(class_name)
     }
 
     pub(super) fn force_lazy_list(&mut self, list: &LazyList) -> Result<Vec<Value>, RuntimeError> {
